+++ src/main/java/com/team13/teamplay2insta/dto/comment/CommentDeleteRequestDto.java
@@ -1,0 +1,8 @@
+package com.team13.teamplay2insta.dto.comment;
+
+import lombok.Data;
+
+@Data
+public class CommentDeleteRequestDto {
+    Long commentid; //api가 소문자...
+}
