+++ src/main/java/com/team13/teamplay2insta/dto/comment/CommentResponseDto.java
@@ -1,4 +1,4 @@
-package com.team13.teamplay2insta.dto;
+package com.team13.teamplay2insta.dto.comment;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
