+++ src/main/java/com/team13/teamplay2insta/dto/user/LoginRequestDto.java
@@ -1,6 +1,5 @@
-package com.team13.teamplay2insta.dto.User;
+package com.team13.teamplay2insta.dto.user;
 
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -8,17 +7,14 @@ import lombok.Setter;
 @Getter
 @Setter
 @NoArgsConstructor
-public class SignupRequestDto {
+public class LoginRequestDto {
 
     private String username; //닉네임
 
-    private String name;//실명
-
     private String pwd;
 
-    public SignupRequestDto(String username, String name, String pwd){
+    public LoginRequestDto(String username, String pwd){
         this.username = username;
-        this.name = name;
         this.pwd = pwd;
     }
 }
