+++ src/main/java/com/team13/teamplay2insta/dto/user/SignupRequestDto.java
@@ -1,4 +1,4 @@
-package com.team13.teamplay2insta.dto.User;
+package com.team13.teamplay2insta.dto.user;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -7,14 +7,17 @@ import lombok.Setter;
 @Getter
 @Setter
 @NoArgsConstructor
-public class LoginRequestDto {
+public class SignupRequestDto {
 
     private String username; //닉네임
 
+    private String name;//실명
+
     private String pwd;
 
-    public LoginRequestDto(String username, String pwd){
+    public SignupRequestDto(String username, String name, String pwd){
         this.username = username;
+        this.name = name;
         this.pwd = pwd;
     }
 }
